refactor: MovieRenderContract 를 utils 패키지로 이동
